Fix gh#76 share rate limit DB loaders
