+++ .clawpatch/features/feat_gh-76_f5b6620cc4.json
@@ -1,1 +1,99 @@
-{"schemaVersion":1,"featureId":"feat_gh-76_f5b6620cc4","title":"gh#76: R6.1 — Push rate-limit DB loaders to @sms/shared/rate-limit/loaders.ts","summary":"Imported from GitHub issue gh#76 in BACKLOG.md section \"Candidate 6 — Rate-limit loader dedupe\".","kind":"job","source":"github-issue-import","confidence":"high","entrypoints":[{"path":"packages/api/src/index.ts","symbol":null,"route":null,"command":null},{"path":"packages/shared/src/rate-limit/check-budget.ts","symbol":null,"route":null,"command":null},{"path":"packages/shared/src/rate-limit/loaders.ts","symbol":null,"route":null,"command":null}],"ownedFiles":[{"path":"packages/api/src/index.ts","reason":"mentioned by issue body"},{"path":"packages/shared/src/rate-limit/check-budget.ts","reason":"mentioned by issue body"},{"path":"packages/shared/src/rate-limit/loaders.ts","reason":"mentioned by issue body"},{"path":"packages/shared/src/package.json","reason":"mentioned by issue body"},{"path":"packages/api/src/services/rate-limit.service.ts","reason":"mentioned by issue body"},{"path":"packages/worker/src/rate-limit.ts","reason":"mentioned by issue body"}],"contextFiles":[{"path":"docs/adr/0007-drizzle-in-shared-allowed-for-io-utilities-not-pure-logic.md","reason":"mentioned by issue body"},{"path":"BACKLOG.md","reason":"backlog queue item for gh#76"}],"tests":[],"tags":["github-issue","gh#76","github-url:https://github.com/joshgk00/social-media-scheduler/issues/76","backlog-order:0021","backlog-section:candidate-6-rate-limit-loader-dedupe","label:enhancement"],"trustBoundaries":["network","filesystem","database","concurrency","external-api"],"status":"needs-fix","lock":null,"findingIds":["fnd_sig-gh-76-c77755017c_28c671c944"],"patchAttemptIds":[],"analysisHistory":[{"runId":"20260519T201935-c24f6a","kind":"github-issue-import","summary":"Imported GitHub issue gh#76","provider":"gh","model":null,"createdAt":"2026-05-19T20:19:35.980Z"}],"createdAt":"2026-05-19T20:19:35.980Z","updatedAt":"2026-05-19T20:19:35.980Z"}
+{
+  "schemaVersion": 1,
+  "featureId": "feat_gh-76_f5b6620cc4",
+  "title": "gh#76: R6.1 — Push rate-limit DB loaders to @sms/shared/rate-limit/loaders.ts",
+  "summary": "Imported from GitHub issue gh#76 in BACKLOG.md section \"Candidate 6 — Rate-limit loader dedupe\".",
+  "kind": "job",
+  "source": "github-issue-import",
+  "confidence": "high",
+  "entrypoints": [
+    {
+      "path": "packages/api/src/index.ts",
+      "symbol": null,
+      "route": null,
+      "command": null
+    },
+    {
+      "path": "packages/shared/src/rate-limit/check-budget.ts",
+      "symbol": null,
+      "route": null,
+      "command": null
+    },
+    {
+      "path": "packages/shared/src/rate-limit/loaders.ts",
+      "symbol": null,
+      "route": null,
+      "command": null
+    }
+  ],
+  "ownedFiles": [
+    {
+      "path": "packages/api/src/index.ts",
+      "reason": "mentioned by issue body"
+    },
+    {
+      "path": "packages/shared/src/rate-limit/check-budget.ts",
+      "reason": "mentioned by issue body"
+    },
+    {
+      "path": "packages/shared/src/rate-limit/loaders.ts",
+      "reason": "mentioned by issue body"
+    },
+    {
+      "path": "packages/shared/src/package.json",
+      "reason": "mentioned by issue body"
+    },
+    {
+      "path": "packages/api/src/services/rate-limit.service.ts",
+      "reason": "mentioned by issue body"
+    },
+    {
+      "path": "packages/worker/src/rate-limit.ts",
+      "reason": "mentioned by issue body"
+    }
+  ],
+  "contextFiles": [
+    {
+      "path": "docs/adr/0007-drizzle-in-shared-allowed-for-io-utilities-not-pure-logic.md",
+      "reason": "mentioned by issue body"
+    },
+    {
+      "path": "BACKLOG.md",
+      "reason": "backlog queue item for gh#76"
+    }
+  ],
+  "tests": [],
+  "tags": [
+    "github-issue",
+    "gh#76",
+    "github-url:https://github.com/joshgk00/social-media-scheduler/issues/76",
+    "backlog-order:0021",
+    "backlog-section:candidate-6-rate-limit-loader-dedupe",
+    "label:enhancement"
+  ],
+  "trustBoundaries": [
+    "network",
+    "filesystem",
+    "database",
+    "concurrency",
+    "external-api"
+  ],
+  "status": "fixed",
+  "lock": null,
+  "findingIds": [
+    "fnd_sig-gh-76-c77755017c_28c671c944"
+  ],
+  "patchAttemptIds": [],
+  "analysisHistory": [
+    {
+      "runId": "20260519T201935-c24f6a",
+      "kind": "github-issue-import",
+      "summary": "Imported GitHub issue gh#76",
+      "provider": "gh",
+      "model": null,
+      "createdAt": "2026-05-19T20:19:35.980Z"
+    }
+  ],
+  "createdAt": "2026-05-19T20:19:35.980Z",
+  "updatedAt": "2026-05-20T21:36:13.845Z"
+}
